extract theme from index file
