+++ src/assets/theme/index.ts
@@ -15,65 +15,4 @@
  */
 
 export * from './helpers';
-
-/*
-=========================================================
-* Material Kit 2 React - v2.0.0
-=========================================================
-
-* Product Page: https://www.creative-tim.com/product/material-kit-react
-* Copyright 2021 Creative Tim (https://www.creative-tim.com)
-
-Coded by www.creative-tim.com
-
- =========================================================
-
-* The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software.
-*/
-
-// @mui material components
-import { createTheme, Theme } from '@mui/material';
-
-import { responsiveFontSizes } from './responsiveFont';
-
-// Material Kit 2 React base styles
-import { palette, breakpoints, typography, globals } from './base';
-
-// Material Kit 2 React components base styles for @mui material components
-import {
-  MuiIconButton,
-  MuiTooltip,
-  MuiAppBar,
-  MuiContainer,
-  MuiIcon,
-  MuiSvgIcon,
-  MuiLink,
-  MuiCard,
-} from './components';
-
-const generatedTheme = createTheme({
-  breakpoints,
-  palette,
-  typography,
-  components: {
-    MuiCssBaseline: {
-      styleOverrides: {
-        globals,
-      },
-    },
-    MuiIconButton,
-    MuiTooltip,
-    MuiAppBar,
-    MuiIcon,
-    MuiSvgIcon,
-    MuiLink,
-    MuiContainer,
-    MuiCard,
-  },
-});
-
-export const theme: Theme =
-  //Generate responsive typography settings based on the options received.
-  responsiveFontSizes(generatedTheme, {
-    breakpoints: generatedTheme.breakpoints.keys,
-  });
+export { theme } from './theme';
